refac: remove interactionHandler (old impl)
