+++ src/App.tsx
@@ -8,35 +8,41 @@ import Controls from './components/Controls';
 import useSelectMode from './hooks/mode/useSelectMode';
 import { useViewport } from './hooks/canvas/useViewport';
 import { useCanvasRenderer } from './hooks/canvas/useCanvasRenderer';
+import { useEventBus } from './hooks/canvas/useEventBus';
+
+const viewportSettings = { // TODO: add this to a configuration file
+  initialX: 0,
+  initialY: 0,
+  initialScale: 1,
+  minScale: 0.1,
+  maxScale: 10,
+};
 
 export default function App() {
   const canvasRef = useRef<HTMLCanvasElement>(null);
-  const { mode, setMode } = useModeManager(useSelectMode());
-  const viewportSettings = { // TODO: add this to a configuration file
-    initialX: 0,
-    initialY: 0,
-    initialScale: 1,
-    minScale: 0.1,
-    maxScale: 10,
-  }
-  const { viewport, handlePan, handleZoom } = useViewport(...Object.values(viewportSettings));
+  const eventBus = useEventBus();
 
-  const render = useCallback((ctx: CanvasRenderingContext2D) => {
-    ctx.beginPath();
-    ctx.arc(100, 100, 50, 0, 2 * Math.PI);
-    ctx.strokeStyle = 'white';
-    ctx.stroke();
-  }, []);
+  const modeManager = useModeManager(useSelectMode());
+  const viewport = useViewport(...Object.values(viewportSettings));
 
-  const { draw } = useCanvasRenderer(canvasRef, viewport, render);
-  useEventAttachment(mode, handlePan, handleZoom);
-  useCanvasInteraction(canvasRef);
+  useCanvasInteraction(canvasRef, eventBus);
+
+  // const render = useCallback((ctx: CanvasRenderingContext2D) => {
+  //   ctx.beginPath();
+  //   ctx.arc(100, 100, 50, 0, 2 * Math.PI);
+  //   ctx.strokeStyle = 'white';
+  //   ctx.stroke();
+  // }, []);
+
+  // useCanvasRenderer(canvasRef, viewport, render);
+  // useEventAttachment(mode, handlePan, handleZoom);
+  // useCanvasInteraction(canvasRef);
 
   return (
     <div>
-      <p>Current mode: {mode.name}</p>
+      {/* <p>Current mode: {mode.name}</p>
       <Controls setMode={setMode} />
-      <Canvas ref={canvasRef} />
+      <Canvas ref={canvasRef} /> */}
     </div>
   );
 }
