+++ src/hooks/mode/useModeManager.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Mode } from '../../types/Mode';
+import { Mode } from '../../types';
 
 export function useModeManager(initialMode: Mode) {
   const [mode, setMode] = useState<Mode>(initialMode);
