+++ src/hooks/mode/useSelectMode.ts
@@ -1,21 +1,21 @@
-import { Mode } from "../../types/Mode";
+import { Mode } from "../../types";
 
 const useSelectMode = (): Mode => {
-    return {
-        name: "Select",
-        handleMouseMove : (e: MouseEvent) => {
-            console.log("Mouse moved at", e.clientX, e.clientY);
-        },
-        handleDragging: (e: MouseEvent) => {
-            console.log("Dragging at", e.clientX, e.clientY);
-        },
-        handleDragEnd: (e: MouseEvent) => {
-            console.log("Selection ended at", e.clientX, e.clientY);
-        },
-        handleClick : (e: MouseEvent) => {
-            console.log("Click at", e.clientX, e.clientY);
-        },
-    };
+  return {
+    name: "Select",
+    handleMouseMove: (e: MouseEvent) => {
+      console.log("Mouse moved at", e.clientX, e.clientY);
+    },
+    handleDragging: (e: MouseEvent) => {
+      console.log("Dragging at", e.clientX, e.clientY);
+    },
+    handleDragEnd: (e: MouseEvent) => {
+      console.log("Selection ended at", e.clientX, e.clientY);
+    },
+    handleClick: (e: MouseEvent) => {
+      console.log("Click at", e.clientX, e.clientY);
+    },
+  };
 };
 
 export default useSelectMode;
