+++ src/main.tsx
@@ -2,12 +2,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
-import { InteractionProvider } from './contexts/InteractionContext'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <InteractionProvider>
-      <App />
-    </InteractionProvider>
+    <App />
   </StrictMode>,
 )
