--- src/types/Mode.ts
@@ -1,7 +1,0 @@
-export interface Mode {
-    name: string;
-    handleDragging(e: MouseEvent): void;
-    handleDragEnd(e: MouseEvent): void;
-    handleClick(e: MouseEvent): void;
-    handleMouseMove(e: MouseEvent): void;
-}
